+++ Deltos/GlobalUsings.cs
@@ -9,6 +9,7 @@ global using System.IO;
 global using System.IO.Compression;
 global using System.Linq;
 global using System.Net;
+global using System.Reflection;
 global using System.Text;
 global using System.Text.RegularExpressions;
 global using System.Text.Json.Serialization;